+++ CapaModelo/OrdenVenta.cs
@@ -36,11 +36,11 @@ namespace CapaModelo
         public int IdProducto { get; set; }
         public string Codigo { get; set; }
         public string NombreProducto { get; set; }
-        public int Cantidad { get; set; }
+        public decimal Cantidad { get; set; }
         public decimal PrecioUnidad { get; set; }
         public decimal IvaPorcentaje { get; set; }
         public decimal TotalLinea { get; set; }
         public decimal TotalLineaIva { get; set; }
-        public int StockDisponible { get; set; }
+        public decimal StockDisponible { get; set; }
     }
 }
